% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgsmbiostats-package.R
\docType{package}
\name{vgsmbiostats-package}
\alias{vgsmbiostats}
\alias{vgsmbiostats-package}
\title{vgsmbiostats: Data from Regression Methods in Biostatistics (2nd Ed.)}
\description{
Datasets from Regression Methods in Biostatistics: Linear, Logistic, Survival, and Repeated Measures Models (2nd Ed., 2012) by Vittinghoff, Glidden, Shiboski, McCulloch. Book website: https://regression.ucsf.edu/. For some reason, I had trouble accessing this site when I was studying the book, and had to scour GitHub to find the accompanying datasets. To "preserve" the datasets for my current and future study, I've put it together in this package.
}
\author{
\strong{Maintainer}: Mirza Khan \email{mirza.khan@vumc.org} (\href{https://orcid.org/0000-0001-7007-9437}{ORCID})

}
\keyword{internal}
